move convertToWorkspace
