+++ apps/vs-code-designer/src/app/commands/convertToWorkspace.ts
@@ -1,17 +1,17 @@
-import { extensionCommand } from '../../../../constants';
-import { localize } from '../../../../localize';
-import { addLocalFuncTelemetry } from '../../../utils/funcCoreTools/funcVersion';
-import { WorkspaceFolderStep } from './workspaceFolderStep';
-import { OpenFolderStep } from './openFolderStep';
+import { extensionCommand } from '../../constants';
+import { localize } from '../../localize';
+import { addLocalFuncTelemetry } from '../utils/funcCoreTools/funcVersion';
+import { WorkspaceFolderStep } from './createWorkspace/createWorkspaceSteps/workspaceFolderStep';
+import { OpenFolderStep } from './createWorkspace/createWorkspaceSteps/openFolderStep';
 import { AzureWizard, DialogResponses } from '@microsoft/vscode-azext-utils';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 import type { IFunctionWizardContext } from '@microsoft/vscode-extension-logic-apps';
 import * as vscode from 'vscode';
 import { window } from 'vscode';
-import { getWorkspaceFile, getWorkspaceFileInParentDirectory, getWorkspaceFolder, getWorkspaceRoot } from '../../../utils/workspace';
-import { WorkspaceNameStep } from './workspaceNameStep';
-import { WorkspaceFileStep } from './workspaceFileStep';
-import { isLogicAppProjectInRoot } from '../../../utils/verifyIsProject';
+import { getWorkspaceFile, getWorkspaceFileInParentDirectory, getWorkspaceFolder, getWorkspaceRoot } from '../utils/workspace';
+import { WorkspaceNameStep } from './createWorkspace/createWorkspaceSteps/workspaceNameStep';
+import { WorkspaceFileStep } from './createWorkspace/createWorkspaceSteps/workspaceFileStep';
+import { isLogicAppProjectInRoot } from '../utils/verifyIsProject';
 
 export async function convertToWorkspace(context: IActionContext): Promise<boolean> {
   const workspaceFolder = await getWorkspaceFolder(context, undefined, true);
