+++ apps/vs-code-designer/src/app/commands/createCustomCodeFunction/createCustomCodeFunction.ts
@@ -5,7 +5,7 @@
 import { ExistingWorkspaceStep } from '../createProject/createProjectSteps/existingWorkspaceStep';
 import { isString } from '@microsoft/logic-apps-shared';
 import { type IFunctionWizardContext, ProjectType } from '@microsoft/vscode-extension-logic-apps';
-import { convertToWorkspace } from '../createWorkspace/createWorkspaceSteps/convertToWorkspace';
+import { convertToWorkspace } from '../convertToWorkspace';
 import { addLocalFuncTelemetry } from '../../utils/funcCoreTools/funcVersion';
 import { type IActionContext, AzureWizard, UserCancelledError } from '@microsoft/vscode-azext-utils';
 import { localize } from '../../../localize';
