+++ apps/vs-code-designer/src/app/commands/generateDeploymentScripts/generateDeploymentScripts.ts
@@ -17,7 +17,7 @@ import { getWorkspaceFolder, isMultiRootWorkspace } from '../../utils/workspace'
 import { AzureWizard, UserCancelledError, type IActionContext } from '@microsoft/vscode-azext-utils';
 import type { IProjectWizardContext } from '@microsoft/vscode-extension-logic-apps';
 import { DeploymentScriptTypeStep } from './generateDeploymentScriptsSteps/DeploymentScriptTypeStep';
-import { convertToWorkspace } from '../createWorkspace/createWorkspaceSteps/convertToWorkspace';
+import { convertToWorkspace } from '../convertToWorkspace';
 import type { SlotTreeItem } from '../../tree/slotsTree/SlotTreeItem';
 import type * as vscode from 'vscode';
 import * as path from 'path';
